refactor(core): share type-resolution between trace and per-type passes
The trace target-type pass and the per-type attribute pass both need
to resolve an entry's effective core type via the same chain
(explicit Type → profile-classified entry.type → URI-scheme
inference for Reference shape). Extract the helper into
`validator/type_resolution.ts` so both passes get the URI-scheme
inference branch landed in 528f6c9.

The observable effect is that MSL-T022 now fires on Reference-shape
entries with no explicit `Type:` when an attribute is incompatible
with the inferred type — e.g., `Bus-protocol` on a `pkg:cargo/*`
Reference fires MSL-T022 because pkg:cargo infers SoftwareComponent
and Bus-protocol is HardwareInterface-only.

- core/validator/type_resolution.ts: new module with shared
  explicitType() and resolvedCoreType() helpers.
- core/validator/trace_types.ts: drop local copies, import from
  the shared module.
- core/validator/per_type_attrs.ts: same — replace explicitType()
  lookup with resolvedCoreType() so URI-scheme inference flows
  through to MSL-T022.
- tests/e2e/validate_test.ts: cover
    pkg:cargo/* Reference + Bus-protocol → MSL-T022 (warning).

Second slice of the Reference-shape PR (prompt-01-ref).
