+++ packages/markspec/core/validator/per_type_attrs.ts
@@ -2,20 +2,22 @@
  * @module core/validator/per_type_attrs
  *
  * Per-type attribute compatibility check (spec §4.3 — MSL-T022). When
- * an entry carries an explicit `Type:` that resolves to a core type,
- * any attribute that's known to the core hierarchy but not allowed on
- * that type fires MSL-T022 as a warning. Unknown attributes (neither
- * universal nor core-typed) fall through to the existing MSL-R010
- * unknown-attribute warning in `validator/mod.ts`.
+ * an entry's effective type resolves to a core type — via explicit
+ * `Type:`, profile classification, or URI-scheme inference for
+ * Reference-shape entries — any attribute that's known to the core
+ * hierarchy but not allowed on that type fires MSL-T022 as a warning.
+ * Unknown attributes (neither universal nor core-typed) fall through
+ * to the existing MSL-R010 unknown-attribute warning in
+ * `validator/mod.ts`.
  */
 
 import type { Diagnostic, Entry } from "../model/mod.ts";
 import {
   attributesForType,
-  CORE_TYPE_HIERARCHY,
   CORE_TYPE_SCOPED_ATTRS,
   UNIVERSAL_ATTRIBUTE_KEYS,
 } from "../model/mod.ts";
+import { resolvedCoreType } from "./type_resolution.ts";
 
 /**
  * Attribute keys that are universal to every entry regardless of type
@@ -33,28 +35,19 @@ const UNIVERSAL_KEYS: ReadonlySet<string> = new Set<string>([
   "Reference-document",
 ]);
 
-/** Find the explicit `Type:` attribute on an entry. */
-function explicitType(entry: Entry): string | undefined {
-  for (const attr of entry.rawAttributes) {
-    if (attr.key === "Type") return attr.value.trim();
-  }
-  return undefined;
-}
-
 /**
  * Emit MSL-T022 warnings for attributes that are core-known but not
- * valid on the entry's resolved type. Skips entries without an
- * explicit `Type:`, entries whose `Type:` is not a core type (those
- * are handled by MSL-T020 / MSL-T023 / MSL-T021), and attributes that
- * are universal or unknown to the core hierarchy entirely (those go
- * through MSL-R010).
+ * valid on the entry's resolved type. Skips entries whose type cannot
+ * be resolved (no explicit `Type:`, no profile classification, no URI
+ * scheme inference — those are handled by MSL-T020 / MSL-T021 /
+ * MSL-T023) and attributes that are universal or entirely unknown to
+ * the core hierarchy (those go through MSL-R010).
  */
 export function validatePerTypeAttributes(
   entry: Entry,
 ): readonly Diagnostic[] {
-  const type = explicitType(entry);
+  const type = resolvedCoreType(entry);
   if (type === undefined) return [];
-  if (!CORE_TYPE_HIERARCHY[type]) return [];
 
   const allowed = attributesForType(type);
   const diagnostics: Diagnostic[] = [];
